Split Upgrade round segmented control
