+++ podcasts/Onboarding/Plus/UpgradeRoundedSegmentedControl.swift
@@ -1,9 +1,53 @@
-//
-//  UpgradeRoundedSegmentedControl.swift
-//  podcasts
-//
-//  Created by Sérgio Estêvão on 01/02/2024.
-//  Copyright © 2024 Shifty Jelly. All rights reserved.
-//
-
 import Foundation
+import SwiftUI
+
+struct UpgradeRoundedSegmentedControl: View {
+    @Binding private var selected: PlanFrequency
+
+    init(selected: Binding<PlanFrequency>) {
+        self._selected = selected
+    }
+
+    var body: some View {
+        HStack(spacing: 0) {
+            Button(L10n.yearly) {
+                withAnimation {
+                    selected = .yearly
+                }
+            }
+            .buttonStyle(UpgradeSegmentedControlButtonStyle(isSelected: selected == .yearly))
+            .padding(4)
+
+            Button(L10n.monthly) {
+                withAnimation {
+                    selected = .monthly
+                }
+            }
+            .buttonStyle(UpgradeSegmentedControlButtonStyle(isSelected: selected == .monthly))
+            .padding(4)
+        }
+        .background(.white.opacity(0.16))
+        .cornerRadius(24)
+    }
+}
+
+struct UpgradeSegmentedControlButtonStyle: ButtonStyle {
+    let isSelected: Bool
+
+    init(isSelected: Bool = true) {
+        self.isSelected = isSelected
+    }
+
+    func makeBody(configuration: Configuration) -> some View {
+        configuration.label
+            .padding(.horizontal, 16)
+            .padding(.vertical, 8)
+            .background(
+                isSelected ? .white : configuration.isPressed ? .white.opacity(0.1) : .clear
+            )
+            .font(style: .subheadline, weight: .medium)
+            .foregroundColor(isSelected ? .black : .white)
+            .cornerRadius(100)
+            .contentShape(Rectangle())
+    }
+}
